# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CommuteAds Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /device/playlist:
    get:
      operationId: getDevicePlaylist
      tags: [playlist]
      summary: Smart ad pacing playlist for a device
      parameters:
        - in: query
          name: device_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ordered playlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaylistItem"

  # ── Campaigns ──────────────────────────────────────────────────────────────
...
  /dynamic-content:
    get:
      operationId: listDynamicContent
      tags: [dynamic-content]
      summary: List dynamic content items
      parameters:
        - in: query
          name: content_type
          schema:
            type: string
            enum: [WEATHER_CARD, NEWS_CARD, BREAKING_ALERT, MEME]
      responses:
        "200":
          description: List of dynamic content
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DynamicContent"
    post:
      operationId: createDynamicContent
      tags: [dynamic-content]
      summary: Create a dynamic content item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DynamicContentInput"
      responses:
        "201":
          description: Created content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DynamicContent"
  /dynamic-content/{id}:
    get:
      operationId: getDynamicContent
      tags: [dynamic-content]
      summary: Get a content item by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dynamic content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DynamicContent"
        "404":
          description: Not found
    patch:
      operationId: updateDynamicContent
      tags: [dynamic-content]
      summary: Update a dynamic content item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DynamicContentUpdate"
      responses:
        "200":
          description: Updated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DynamicContent"
    delete:
      operationId: deleteDynamicContent
      tags: [dynamic-content]
      summary: Delete a dynamic content item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /dynamic-content/push-breaking-news:
    post:
      operationId: pushBreakingNews
      tags: [dynamic-content]
      summary: Push an instant breaking news alert to all devices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BreakingNewsInput"
      responses:
        "201":
          description: Breaking news pushed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DynamicContent"

  # ── Proof of Plays ─────────────────────────────────────────────────────────
  /proof-of-plays:
    get:
      operationId: listProofOfPlays
      tags: [proof-of-plays]
      summary: List proof of play records
      parameters:
        - in: query
          name: campaign_id
          schema:
            type: integer
        - in: query
          name: device_id
          schema:
            type: integer
      responses:
        "200":
          description: List of proof of plays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProofOfPlay"
    post:
      operationId: createProofOfPlay
      tags: [proof-of-plays]
      summary: Log a proof of play event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProofOfPlayInput"
      responses:
        "201":
          description: Logged proof of play
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProofOfPlay"

  # ── SOS Alerts ─────────────────────────────────────────────────────────────
  /sos:
    post:
      operationId: createSosAlert
      tags: [sos-alerts]
      summary: Submit an SOS emergency alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SosAlertInput"
      responses:
        "201":
          description: SOS alert logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SosAlert"
  /sos-alerts:
    get:
      operationId: listSosAlerts
      tags: [sos-alerts]
      summary: List SOS alerts
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, RESOLVED]
      responses:
        "200":
          description: List of SOS alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SosAlert"
  /sos-alerts/{id}/resolve:
    post:
      operationId: resolveSosAlert
      tags: [sos-alerts]
      summary: Resolve an SOS alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SosAlert"
components:
  schemas:
...
    PlaylistItem:
      type: object
      required: [type, id, title, durationSec, order]
      properties:
        type:
          type: string
          enum: [AD, WEATHER_CARD, NEWS_CARD, BREAKING_ALERT, MEME]
        id:
          type: integer
        title:
          type: string
        videoUrl:
          type: ["string", "null"]
        durationSec:
          type: integer
        order:
          type: integer
        campaignId:
          type: ["integer", "null"]

    # ── Campaign ──────────────────────────────────────────────────────────────
...
    DynamicContent:
      type: object
      required: [id, contentType, headline, bodyText, locationTag, validUntil]
      properties:
        id:
          type: integer
        contentType:
          type: string
          enum: [WEATHER_CARD, NEWS_CARD, BREAKING_ALERT, MEME]
        headline:
          type: string
        bodyText:
          type: string
        iconUrl:
          type: ["string", "null"]
        locationTag:
          type: string
        validUntil:
          type: string
    DynamicContentInput:
      type: object
      required: [contentType, headline, bodyText, locationTag, validUntil]
      properties:
        contentType:
          type: string
          enum: [WEATHER_CARD, NEWS_CARD, BREAKING_ALERT, MEME]
        headline:
          type: string
        bodyText:
          type: string
        iconUrl:
          type: string
        locationTag:
          type: string
        validUntil:
          type: string
    DynamicContentUpdate:
      type: object
      properties:
        headline:
          type: string
        bodyText:
          type: string
        iconUrl:
          type: string
        locationTag:
          type: string
        validUntil:
          type: string
    BreakingNewsInput:
      type: object
      required: [headline, bodyText]
      properties:
        headline:
          type: string
        bodyText:
          type: string
        locationTag:
          type: string

    # ── Proof of Play ─────────────────────────────────────────────────────────
    ProofOfPlay:
      type: object
      required: [id, deviceId, adAssetId, playedAt, latitude, longitude, passengerVerified]
      properties:
        id:
          type: integer
        deviceId:
          type: integer
        adAssetId:
          type: integer
        playedAt:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        passengerVerified:
          type: boolean
    ProofOfPlayInput:
      type: object
      required: [deviceId, adAssetId, latitude, longitude]
      properties:
        deviceId:
          type: integer
        adAssetId:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        passengerVerified:
          type: boolean

    # ── SOS Alert ─────────────────────────────────────────────────────────────
    SosAlert:
      type: object
      required: [id, deviceId, alertType, latitude, longitude, timestamp, status]
      properties:
        id:
          type: integer
        deviceId:
          type: integer
        alertType:
          type: string
          enum: [POLICE, FIRE, AMBULANCE, WOMENS_SAFETY]
        latitude:
          type: number
        longitude:
          type: number
        timestamp:
          type: string
        status:
          type: string
          enum: [PENDING, RESOLVED]
        nodeCode:
          type: ["string", "null"]
    SosAlertInput:
      type: object
      required: [deviceId, alertType, latitude, longitude]
      properties:
        deviceId:
          type: integer
        alertType:
          type: string
          enum: [POLICE, FIRE, AMBULANCE, WOMENS_SAFETY]
        latitude:
          type: number
        longitude:
          type: number