# lib/api-zod/src/generated/types/dynamicContent.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CommuteAds Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DynamicContentContentType } from './dynamicContentContentType';

export interface DynamicContent {
  id: number;
  contentType: DynamicContentContentType;
  headline: string;
  bodyText: string;
  /** @nullable */
  iconUrl?: string | null;
  locationTag: string;
  validUntil: string;
}