# lib/api-zod/src/generated/types/proofOfPlay.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CommuteAds Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProofOfPlay {
  id: number;
  deviceId: number;
  adAssetId: number;
  playedAt: string;
  latitude: number;
  longitude: number;
  passengerVerified: boolean;
}