# lib/api-zod/src/generated/types/proofOfPlayInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CommuteAds Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProofOfPlayInput {
  deviceId: number;
  adAssetId: number;
  latitude: number;
  longitude: number;
  passengerVerified?: boolean;
}