# lib/api-zod/src/generated/types/sosAlertInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CommuteAds Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SosAlertInputAlertType } from './sosAlertInputAlertType';

export interface SosAlertInput {
  deviceId: number;
  alertType: SosAlertInputAlertType;
  latitude: number;
  longitude: number;
}